% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/personograph.R
\name{calc.ier}
\alias{calc.ier}
\title{Calculate the IER (Intervention Event Rates)}
\usage{
calc.ier(cer, point, sm)
}
\arguments{
\item{cer}{Absolute risk with control (calculated; from 0 to 1)}

\item{point}{Relative risk with intervention (direct from meta-analysis)}

\item{sm}{The outcome measure, RR or OR as string}
}
\value{
Absolute risk with intervention as Intervention Event Rates (IER)
}
\description{
Calculate the IER (Intervention Event Rates)
}
\seealso{
\code{\link{w.approx.cer}}
}

