% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/personograph.R
\name{personograph}
\alias{personograph}
\title{Plots a personograph}
\usage{
personograph(data, fig.title = NULL, fig.cap = NULL, draw.legend = T,
  icon = NULL, icon.dim = NULL, icon.style = 1, n.icons = 100,
  plot.width = 0.6, dimensions = ceiling(sqrt(c(n.icons, n.icons))),
  colors = as.colors(data))
}
\arguments{
\item{data}{A list of names to percentages (from 0 to 1)}

\item{fig.title}{Figure title}

\item{fig.cap}{Figure caption}

\item{draw.legend}{Logical indicating whether to draw the legend}

\item{icon}{A \code{grImport} \code{Picture} for the icon, overwrites \code{icon.style}}

\item{icon.dim}{The dimensions of icon as a vector \code{c(width, height)} of \code{unit} or numerical, calculated from the \code{dimensions} if not supplied}

\item{icon.style}{A numeric from 1-11 indicating which of the included icons to use}

\item{n.icons}{Number of icons to draw, defaults to 100}

\item{plot.width}{The percentage of width that the main plotting area should take (with respect to the frame)}

\item{dimensions}{A vector of \code{c(rows, columns)} for the dimensions of the grid}

\item{colors}{A vector of names to colors, must match the names in data. Uses \code{gray.colors} style if none supplied}
}
\value{
None.
}
\description{
Plots a personograph from a list with with percentages (must sum to
1). A personograph is a graphical represenation of relative benefit
or harm, using a grid of icons with different colors. Its intended
use is similar to that of Cates Plots (Visual Rx, Number Needed to
Treat visualization).
Although these could be seen as Kuiper-Marshall plots.
}
\examples{
data <- list(good= 0.8884758, benefit = 0.04784283, harmed = 0, bad = 0.06368133)
personograph(data)
}

