% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/personograph.R
\name{w.approx.cer}
\alias{w.approx.cer}
\title{Calculate the CER (Control Event Rates)}
\usage{
w.approx.cer(ev.ctrl, n.ctrl)
}
\arguments{
\item{ev.ctrl}{vector of event rates in the control group (/arm)}

\item{n.ctrl}{a vector of sample sizes in the control group (/arm)}
}
\value{
the approximated CER
}
\description{
from the data, this is a weighted approximation of absolute
risk with control (calculated; from 0 to 1)
}

