% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/personograph.R
\name{calc.ier}
\alias{calc.ier}
\title{Calculate the IER (Intervention Event Rates)}
\usage{
calc.ier(cer, point, units = sm)
}
\arguments{
\item{cer}{absolute risk with control (calculated; from 0 to 1)}

\item{point}{relative risk with intervention (direct from meta-analysis)}

\item{units}{RR or OR as string}
}
\value{
absolute risk with intervention
}
\description{
Calculate the IER (Intervention Event Rates)
}

