% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/personograph.R
\name{uplift}
\alias{uplift}
\title{"Uplift" from IER and CER}
\usage{
uplift(ier, cer, higher_is_better = NULL)
}
\arguments{
\item{ier}{Intervention Event Rates}

\item{cer}{Control Event Rates}

\item{higher_is_better}{bool indicating the direction of the outcome measure, default TRUE}
}
\value{
A list of S3 class \code{personograph.uplift} with the following elements:
\itemize{
\item{happy}{people who are happy no matter what treatment}
\item{sad}{people who are sad no matter what treatment}
\item{helped}{people who would be helped by treatment}
\item{harmed}{people who would be harmed by treatment}
}
}
\description{
Calculates the percentage (from 0 to 1) of people helped, harmed, sad, and happy
from the Intervention Event Rates (IER) and Control Event Rates (CER).
Note that the result depends on the direction of the outcome measure,
e.g. higher_is_better = T (default) for treatment efficacy, higher_is_better = F for
adverse events.
}
\details{
The adopted terminology is similar to that of Uplift modelling
https://en.wikipedia.org/wiki/Uplift_modelling
}
\examples{
ier <- 0.06368133
cer <- 0.1115242
uplift(ier, cer, T)
}

