% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/personograph.R
\name{w.approx.cer}
\alias{w.approx.cer}
\title{Calculate the CER (Control Event Rates)}
\usage{
w.approx.cer(ev.ctrl, n.ctrl)
}
\arguments{
\item{ev.ctrl}{Vector of event rates in the control group (/arm)}

\item{n.ctrl}{A vector of sample sizes in the control group (/arm)}
}
\value{
Approximated Control Event Rates (CER)
}
\description{
Calculates the CER from the data, this is a weighted approximation of absolute
risk with control (from 0 to 1)
}

